add VELOCITY_SUPPLEMENTAL_CONTEXT header
Add a header to specify a Map of supplementary variables to the
default velocity context entries.  Entries in this map will override
the default pre-defined values when using the same key, e.g. "body".

This provides a simpler way to pre-defined variables than providing
a custom context that completely replaces the default, while still
allowing values to be "returned" from the template via headers.