LOG instance removed. Parent Logger is going to be used.
Logging debug level instead of info. 

CMProducer is a Processor. process() method throws Exception. Let camel
do the work and not catch exceptions. 

In order to obtain the payload, Message.getMandatoryBody() is used
instead of getBody() so there is no need to the check if the payload is
null. Camel 's InvalidPayloadException is going to be thrown.